fix small

+++ TlmCmd/block_command_loader.h
@@ -139,14 +139,14 @@ void BCL_tool_prepare_param_float(float val);
  * @note   BCL_tool_register_cmd前でのみ使用することを想定している
  * @param  val        追加するパラメータ
  */
-void BCL_tool_prepare_param_uint64(uint32_t val);
+void BCL_tool_prepare_param_uint64(uint64_t val);
 
 /**
  * @brief  コマンド登録前にコマンドの引数としてint64のパラメータを追加する
  * @note   BCL_tool_register_cmd前でのみ使用することを想定している
  * @param  val        追加するパラメータ
  */
-void BCL_tool_prepare_param_int64(int32_t val);
+void BCL_tool_prepare_param_int64(int64_t val);
 
 /**
  * @brief  コマンド登録前にコマンドの引数としてdoubleのパラメータを追加する
